Remove unnecessary suppressions
