+++ test/EliteChroma.Core.Tests/Internal/Event.cs
@@ -1,0 +1,18 @@
+﻿namespace EliteChroma.Core.Tests.Internal
+{
+    internal sealed class Event
+    {
+        public Event(bool isStatus, string json, bool changesGameState)
+        {
+            IsStatus = isStatus;
+            Json = json;
+            ChangesGameState = changesGameState;
+        }
+
+        public bool IsStatus { get; }
+
+        public string Json { get; }
+
+        public bool ChangesGameState { get; }
+    }
+}
